fix: avoid breaking tekton reconciliation
on old tekton v0.11.3 (used in jx v2), the pipelinerun reconciliation might fail because the tekton controller can't update the pipelinerun, because it has been changed (by our patching operation).
this is turn result in multiple taskruns being created (1 for each reconciliation attempt).

the "fix" is to wait until the initial reconciliation is finished, which is after the tekton controller has updated the PR labels, adding the `tekton.dev/pipeline` label.
which is why here we are now waiting until this label is present before handling the PR.